Adds a bunch of content, style, and JS
I should have done this staged better, but ¯\_(ツ)_/¯

This does a bunch of things. Instead of making the .js modules with both
code and data, I seperated it out to yaml files for easy reading. This
did involve making the .js modules able to read the data in a yaml
file. This did allow me to make a master module called page.js
that does all the magic. It is easlily extendable to be able to
add different parameters too. This, combined with the data,
makes the module for each of the pages.

I also implemented some things to the config file for webpack
like the needed yaml, image, and css readers, but also creating
the base html file too. It cleans out the dist file before writing
everything.